requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
blinker==1.4
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
command-not-found==0.3
cryptography==41.0.3
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-mjml==1.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.12.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
keyring==23.5.0
kombu==5.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.4.4
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg==3.1.10
pycparser==2.21
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
systemd-python==234
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.6
wrapt==1.15.0
zipp==1.0.0
